Move sensor backends to their own submodule hierarchy
There are enough other miscellaneous infrastructure modules floating
around at the top level (sysfs, gpio, dbus_helpers, etc.) that having
the sensor backends all grouped together seems like a nice improvement.
